Fix compiler warnings with MSC

+++ src/base/strbuf.c
@@ -172,7 +172,7 @@ sg_strbuf_vprintf(struct sg_strbuf *b, const char *fmt, va_list ap)
     rem = b->e - b->p;
     len = _vscprintf(fmt, ap);
     len = len < 0 ? 0 : len;
-    if (len < rem)
+    if ((size_t) len < rem)
         sg_strbuf_reserve(b, len);
     _vsnprintf(b->p, rem, fmt, ap);
     b->p += len;
